// apps/main/src/models/v2/controlplane/runnerRuleSummary.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 */

/**
 * A summary of a rule that matches resource requests against a module
 */
export interface RunnerRuleSummary {
  /** The Organization ID. */
  org_id: string;
  /** The unique identifier for the rule under the organization. */
  id: string;
  /** The unique identifier for a runner */
  runner_id: string;
  /** The time this rule was created */
  created_at: string;
  /** The optional project id that this rule matches. */
  project_id: string;
  /** The optional env type id that this rule matches. */
  env_type_id: string;
}